elf: use snapped strip (#2051)
Currently Snapcraft uses the `strip` utility to remove the
.note.go.buildid section from ELF files if necessary. However, there are
a few issues with the implementation:

  1. The Debian package doesn't depend upon binutils (which contains
     `strip`).

  2. The snap doesn't bundle `strip`, thereby relying on its presence on
     the host.

Binutils isn't always available (e.g. in the Docker images), and snaps
should bundle all their dependencies. Fix this issue by adding binutils
as a dependency in debian/control and as a `stage-package`, and use the
snapped version if running as a snap.

Signed-off-by: Kyle Fazzari <kyrofa@ubuntu.com>